requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=4.9.0
lxml-html-clean>=0.1.0
sqlite3
flask>=3.0.0
python-dateutil>=2.8.2
